requirements.txt
flask==2.3.3
flask-cors==4.0.0
deep-translator==1.11.4
python-dotenv==1.0.0
requests==2.31.0 